gRPC/requirements.txt
grpcio==1.68.1
grpcio-tools
mysql-connector-python
yfinance
cachetools
confluent-kafka
prometheus-client